Adds separate "Custom fields" integration test suits
