+++ cypress/integration/custom-fields/index.js
@@ -1,0 +1,3 @@
+describe('Custom fields', function() {
+  require('./BirthDate')
+})
