+++ cypress/integration/index.js
@@ -4,3 +4,4 @@ require('./behavior')
 require('./field-grouping')
 require('./validation')
 require('./reactive-props')
+require('./custom-fields')
